+++ test/test_methods/test_hg.py
@@ -17,7 +17,7 @@ needs_hg = pytest.mark.skipif(
 
 BUILD_DATE = datetime(2038, 1, 19, 3, 14, 7, tzinfo=timezone.utc)
 
-DATA_DIR = Path(__file__).with_name("data")
+DATA_DIR = Path(__file__).parent.with_name("data")
 
 
 @needs_hg